requirements.txt
numpy==2.1.1
sentence_transformers==3.0.1
streamlit==1.38.0
google-generativeai